deleteSchema deals with empty tables/sequences, prepare delete savepoints
git-svn-id: https://svn.exedio.com/cope/cope/trunk@26046 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ runtime/postgresqlsrc/com/exedio/cope/PostgresqlDialect.java
@@ -246,23 +246,27 @@ final class PostgresqlDialect extends Dialect
 	{
 		final StringBuilder bf = new StringBuilder();
 
-		bf.append("truncate ");
-		boolean first = true;
-		for(final Table table : tables)
+		if(!tables.isEmpty())
 		{
-			if(first)
-				first = false;
-			else
-				bf.append(',');
+			bf.append("truncate ");
+			boolean first = true;
+			for(final Table table : tables)
+			{
+				if(first)
+					first = false;
+				else
+					bf.append(',');
 
-			bf.append(table.quotedID);
+				bf.append(table.quotedID);
+			}
+			bf.append(';');
 		}
-		bf.append(';');
 
 		for(final SequenceX sequence : sequences)
 			sequence.delete(bf, this);
 
-		execute(connectionPool, bf.toString());
+		if(bf.length()>0)
+			execute(connectionPool, bf.toString());
 	}
 
 	private static void execute(final ConnectionPool connectionPool, final String sql)
